feat: Add responsible engineer to construction work
Introduces a new field `responsibleId` to the `ConstructionWork` interface to track the technical supervisor.
This change also updates the `WorkOverview` component to allow assigning a responsible engineer and the `ClientDashboard` to display their information.
Icons related to budgeting and responsibility have been added to `FinanceView` and `WorkOverview` respectively.
The `App.tsx` was updated to pass the `users` list to `ClientDashboard`.